backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
pydantic==2.5.2
python-dotenv==1.0.0
websockets==12.0
voximplant-apiclient==1.9.0
python-multipart==0.0.6
